refactor(lib): reduce lib.rs to pure exports facade
- lib.rs: 1296 → 145 lines (89% reduction, module declarations + re-exports only)
- config.rs: ScraperConfig + ConcurrencyConfig + OutputFormat (moved from lib.rs)
- cli/args.rs: Args + Commands + Shell (moved from lib.rs)
- url_validation.rs: validate_and_parse_url (moved from lib.rs)
- cli/mod.rs: updated to include new args module
- 440/440 tests passing, --features ai compiles clean
