+++ src/cli/mod.rs
@@ -1,8 +1,11 @@
-//! CLI module — argument parsing, error handling, completions, config
+//! CLI module — argument parsing, error handling, completions, config.
 //!
 //! Clean Architecture Adapters layer: all CLI-related utilities.
 
+pub mod args;
 pub mod completions;
 pub mod config;
 pub mod error;
 pub mod summary;
+
+pub use args::{Args, Commands, Shell};
