+++ src/lib.rs
@@ -1,28 +1,9 @@
 //! Rust Scraper — Production-ready web scraper with Clean Architecture
 //!
 //! **Rust Scraper** is a high-performance, async web scraper designed for
-//! building RAG (Retrieval-Augmented Generation) datasets. Built with Clean Architecture
-//! principles for production use.
-//!
-//! # Features
-//!
-//! - **Async Web Scraping**: Multi-threaded with Tokio runtime
-//! - **Sitemap Support**: Zero-allocation streaming parser (quick-xml)
-//!   - Gzip decompression (async-compression)
-//!   - Sitemap index recursion (max depth 3)
-//!   - Auto-discovery from `robots.txt`
-//! - **TUI Interactivo**: Ratatui + crossterm URL selector
-//!   - Interactive checkbox selection
-//!   - Confirmation mode before download
-//!   - Terminal restore on panic/exit
-//! - **Clean Architecture**: Domain → Application → Infrastructure → Adapters
-//! - **Error Handling**: `thiserror` for libraries, `anyhow` for applications
-//! - **Performance**: True streaming (~8KB RAM), LazyLock cache, bounded concurrency
-//! - **Security**: SSRF prevention, Windows-safe filenames, WAF bypass prevention
+//! building RAG (Retrieval-Augmented Generation) datasets.
 //!
 //! # Architecture
-//!
-//! Following Clean Architecture with four layers:
 //!
 //! ```text
 //! Domain (entities, errors)
@@ -38,128 +19,56 @@
 //!
 //! # Examples
 //!
-//! ## Basic Usage
-//!
 //! ```no_run
-//! use rust_scraper::{create_http_client, scrape_with_readability, ScraperConfig};
+//! use rust_scraper::{scrape_with_readability, ScraperConfig};
 //!
 //! # #[tokio::main]
 //! # async fn main() -> anyhow::Result<()> {
-//! let client = create_http_client()?;
 //! let url = url::Url::parse("https://example.com")?;
 //! let config = ScraperConfig::default();
-//! let results = scrape_with_readability(&client, &url).await?;
+//! let results = scrape_with_readability(&url, &config).await?;
 //! # Ok(())
 //! # }
 //! ```
-//!
-//! ## URL Discovery with Sitemap
-//!
-//! ```no_run
-//! use rust_scraper::{discover_urls_for_tui, CrawlerConfig};
-//! use url::Url;
-//!
-//! # #[tokio::main]
-//! # async fn main() -> anyhow::Result<()> {
-//! let seed = Url::parse("https://example.com")?;
-//! let config = CrawlerConfig::builder(seed)
-//!     .concurrency(5)
-//!     .use_sitemap(true)
-//!     .build();
-//!
-//! let urls = discover_urls_for_tui("https://example.com", &config).await?;
-//! println!("Found {} URLs", urls.len());
-//! # Ok(())
-//! # }
-//! ```
-//!
-//! ## Custom Configuration
-//!
-//! ```
-//! use rust_scraper::ScraperConfig;
-//!
-//! let config = ScraperConfig::default()
-//!     .with_images()
-//!     .with_documents()
-//!     .with_output_dir("./output".into())
-//!     .with_scraper_concurrency(5);
-//!
-//! assert!(config.download_images);
-//! assert!(config.download_documents);
-//! assert_eq!(config.scraper_concurrency, 5);
-//! ```
-//!
-//! # Error Handling
-//!
-//! This library uses [`thiserror`](https://docs.rs/thiserror) for type-safe error handling.
-//! All fallible functions return [`Result<T, ScraperError>`](Result).
-//!
-//! ```
-//! use rust_scraper::{validate_and_parse_url, ScraperError};
-//!
-//! match validate_and_parse_url("https://example.com") {
-//!     Ok(url) => println!("Valid URL: {}", url),
-//!     Err(ScraperError::InvalidUrl(msg)) => eprintln!("Invalid URL: {}", msg),
-//!     Err(e) => eprintln!("Error: {}", e),
-//! }
-//! ```
-//!
-//! # Performance
-//!
-//! - **Streaming**: Constant ~8KB RAM usage, no OOM risks
-//! - **Zero-Allocation Parsing**: quick-xml for sitemaps
-//! - **LazyLock Cache**: Syntax highlighting (2-10ms → ~0.01ms)
-//! - **Bounded Concurrency**: Configurable parallel downloads
-//!
-//! # Security
-//!
-//! - **SSRF Prevention**: URL host comparison (not string contains)
-//! - **Windows Safe**: Reserved names blocked (`CON` → `CON_safe`)
-//! - **WAF Bypass Prevention**: Chrome 131+ UAs with TTL caching
-//! - **Input Validation**: `url::Url::parse()` (RFC 3986 compliant)
-//!
-//! # Testing
-//!
-//! ```bash
-//! # Run all tests
-//! cargo test
-//!
-//! # Run with output
-//! cargo test -- --nocapture
-//!
-//! # Run specific test
-//! cargo test test_validate_and_parse_url
-//! ```
-//!
-//! **Tests:** 19 passing ✅
-//!
-//! # MSRV
-//!
-//! Minimum Supported Rust Version: 1.75.0
 
 // ============================================================================
-// Lints — rust-skills compliance (CRITICAL: correctness -> HIGH -> MEDIUM)
+// Lints
 // ============================================================================
 #![deny(clippy::correctness)]
 #![warn(clippy::suspicious)]
 #![warn(clippy::style)]
 #![warn(clippy::complexity)]
 #![warn(clippy::perf)]
-#![allow(missing_docs)] // TODO: documentar incrementally
+#![allow(missing_docs)]
 #![warn(clippy::undocumented_unsafe_blocks)]
-#![allow(clippy::module_name_repetitions)] // domain::domain_entity etc.
+#![allow(clippy::module_name_repetitions)]
 
 // ============================================================================
-// Public API Exports
+// Modules
 // ============================================================================
 
 pub mod config;
 pub mod error;
 
-// Domain layer — Core business entities (pure, no dependencies)
 pub mod domain;
 #[cfg(feature = "ai")]
 pub use domain::semantic_cleaner::SemanticCleaner;
+
+pub mod application;
+pub mod infrastructure;
+pub mod adapters;
+pub mod extractor;
+pub mod url_path;
+pub mod user_agent;
+pub mod export_factory;
+pub mod cli;
+pub mod url_validation;
+
+// ============================================================================
+// Re-exports
+// ============================================================================
+
+// Domain layer
 pub use domain::{
     ContentType, CrawlError, CrawlResult, CrawlerConfig, CrawlerConfigBuilder, DiscoveredUrl,
     DownloadedAsset, ExportFormat, JsRenderError, JsRenderer, ScrapedContent, ValidUrl,
@@ -167,8 +76,7 @@ pub use domain::{
 #[cfg(feature = "ai")]
 pub use error::SemanticError;
 
-// Application layer — Use cases (orchestration)
-pub mod application;
+// Application layer
 pub use application::{
     crawl_site, crawl_with_sitemap, create_http_client, detect_spa_content, discover_urls_for_tui,
     extract_domain,
@@ -178,8 +86,7 @@ pub use application::{
     SpaDetectionResult,
 };
 
-// Infrastructure layer — Implementations (technical details)
-pub mod infrastructure;
+// Infrastructure layer
 pub use infrastructure::{
     converter, crawler,
     export::{jsonl_exporter, state_store, vector_exporter},
@@ -188,856 +95,43 @@ pub use infrastructure::{
     scraper::readability,
 };
 
-// Export factory functions
-pub mod export_factory;
+// Adapters
+pub use url_path::{Domain, OutputPath, UrlPath};
+pub use user_agent::{get_random_user_agent_from_pool, UserAgentCache};
 
-// Adapters — External integrations (feature-gated)
-pub mod adapters;
+// Export factory
+pub use export_factory::{create_exporter, domain_from_url, process_results};
 
-// CLI utilities for binary (Adapters layer in Clean Architecture)
-pub mod cli;
+// CLI
 pub use cli::{
     config::{init_logging_dual, is_no_color, should_emit_emoji, ConfigDefaults},
     error::{CliError, CliExit},
     summary::ScrapeSummary,
+    Args, Commands, Shell,
 };
-pub mod extractor;
-pub mod url_path;
-pub mod user_agent;
-pub use url_path::{Domain, OutputPath, UrlPath};
-pub use user_agent::{get_random_user_agent_from_pool, UserAgentCache};
 
-// Public API re-exports (export factory)
-pub use export_factory::{create_exporter, domain_from_url, process_results};
+// Config types
+pub use config::{ConcurrencyConfig, OutputFormat, ScraperConfig};
 
-// CLI types
+// Error and result types
 pub use clap::{Parser, ValueEnum};
 pub use error::{Result, ScraperError};
 
-// Re-export save_results for convenience
+// File saver
 pub use infrastructure::output::file_saver::{save_results, ObsidianOptions};
 
-// T-024: Include build-time metadata generated by built crate
-// This must be at module level because built.rs defines static items
+// URL validation
+pub use url_validation::validate_and_parse_url;
+
+// ============================================================================
+// Build metadata
+// ============================================================================
+
 #[doc(hidden)]
 pub(crate) mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));
 }
 
-// ============================================================================
-// Public Types
-// ============================================================================
-
-/// Output format for scraped content.
-///
-/// # Examples
-///
-/// ```
-/// use rust_scraper::OutputFormat;
-///
-/// let format = OutputFormat::Markdown;
-/// assert_eq!(format, OutputFormat::Markdown);
-/// Output format for scraped content.
-///
-/// # Examples
-///
-/// ```
-/// use rust_scraper::OutputFormat;
-///
-/// let format = OutputFormat::Markdown;
-/// assert_eq!(format, OutputFormat::Markdown);
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
-pub enum OutputFormat {
-    /// Markdown format with YAML frontmatter (recommended for RAG)
-    Markdown,
-    /// Structured JSON with metadata
-    Json,
-    /// Plain text without formatting
-    Text,
-}
-/// use rust_scraper::ScraperConfig;
-///
-/// // Default configuration
-/// let config = ScraperConfig::default();
-///
-/// // Custom configuration with builder pattern
-/// let config = ScraperConfig::default()
-///     .with_images()
-///     .with_documents()
-///     .with_output_dir("./output".into())
-///     .with_scraper_concurrency(5);
-///
-/// assert!(config.download_images);
-/// assert!(config.download_documents);
-/// assert_eq!(config.scraper_concurrency, 5);
-/// ```
-///
-/// # Concurrency Recommendations
-///
-/// | Storage | Concurrency | Reason |
-/// |---------|-------------|--------|
-/// | HDD | 3 (default) | Avoids disk thrashing on mechanical drives |
-/// | SSD | 5-8 | Faster random I/O |
-/// | NVMe | 10+ | Very high IOPS |
-#[derive(Debug, Clone)]
-pub struct ScraperConfig {
-    /// Enable image downloading (PNG, JPG, GIF, WEBP, SVG, BMP)
-    pub download_images: bool,
-    /// Enable document downloading (PDF, DOCX, XLSX, PPTX, etc.)
-    pub download_documents: bool,
-    /// Output directory for downloaded assets
-    pub output_dir: std::path::PathBuf,
-    /// Maximum file size in bytes (default: 50MB)
-    pub max_file_size: Option<u64>,
-    /// Timeout for individual asset downloads in seconds
-    pub download_timeout_secs: u64,
-    /// Maximum concurrent scrapers (default: 3 for HDD-aware on 4C CPU)
-    pub scraper_concurrency: usize,
-}
-
-impl Default for ScraperConfig {
-    fn default() -> Self {
-        Self {
-            download_images: false,
-            download_documents: false,
-            output_dir: std::path::PathBuf::from("output"),
-            max_file_size: Some(50 * 1024 * 1024), // 50MB default
-            download_timeout_secs: 30,
-            scraper_concurrency: 3, // HDD-aware: nproc - 1 for 4C CPU
-        }
-    }
-}
-
-impl ScraperConfig {
-    /// Create a new config with default values.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ScraperConfig;
-    ///
-    /// let config = ScraperConfig::new();
-    /// assert!(!config.download_images);
-    /// ```
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Enable image downloading.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ScraperConfig;
-    ///
-    /// let config = ScraperConfig::default().with_images();
-    /// assert!(config.download_images);
-    /// ```
-    #[must_use]
-    pub fn with_images(mut self) -> Self {
-        self.download_images = true;
-        self
-    }
-
-    /// Enable document downloading.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ScraperConfig;
-    ///
-    /// let config = ScraperConfig::default().with_documents();
-    /// assert!(config.download_documents);
-    /// ```
-    #[must_use]
-    pub fn with_documents(mut self) -> Self {
-        self.download_documents = true;
-        self
-    }
-
-    /// Set custom output directory.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ScraperConfig;
-    ///
-    /// let config = ScraperConfig::default()
-    ///     .with_output_dir("./my-output".into());
-    /// assert_eq!(config.output_dir, std::path::PathBuf::from("./my-output"));
-    /// ```
-    #[must_use]
-    pub fn with_output_dir(mut self, dir: std::path::PathBuf) -> Self {
-        self.output_dir = dir;
-        self
-    }
-
-    /// Set scraper concurrency limit.
-    ///
-    /// # Arguments
-    ///
-    /// * `concurrency` - Maximum concurrent scrapers
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ScraperConfig;
-    ///
-    /// let config = ScraperConfig::default()
-    ///     .with_scraper_concurrency(5);
-    /// assert_eq!(config.scraper_concurrency, 5);
-    /// ```
-    ///
-    /// # Recommendations
-    ///
-    /// - **HDD**: 3 (default) — avoids disk thrashing
-    /// - **SSD**: 5-8 — faster random I/O
-    /// - **NVMe**: 10+ — very high IOPS
-    #[must_use]
-    pub fn with_scraper_concurrency(mut self, concurrency: usize) -> Self {
-        self.scraper_concurrency = concurrency;
-        self
-    }
-
-    /// Check if any download is enabled.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ScraperConfig;
-    ///
-    /// let config = ScraperConfig::default();
-    /// assert!(!config.has_downloads());
-    ///
-    /// let config = config.with_images();
-    /// assert!(config.has_downloads());
-    /// ```
-    pub fn has_downloads(&self) -> bool {
-        self.download_images || self.download_documents
-    }
-}
-
-/// Concurrency configuration with smart auto-detection
-///
-/// Provides intelligent defaults based on hardware capabilities:
-/// - **Auto-detection**: Uses `std::thread::available_parallelism()` to detect CPU cores
-/// - **HDD-aware**: Limits concurrency on systems with limited I/O
-/// - **Safe bounds**: Clamps values between 1 and 16
-///
-/// # Examples
-///
-/// ```
-/// use rust_scraper::ConcurrencyConfig;
-///
-/// // Auto-detect (default)
-/// let config = ConcurrencyConfig::default();
-///
-/// // Explicit value
-/// let config = ConcurrencyConfig::new(5);
-///
-/// // Get the resolved value
-/// let concurrency = config.resolve();
-/// println!("Using {} concurrent workers", concurrency);
-/// ```
-#[derive(Debug, Clone)]
-pub struct ConcurrencyConfig {
-    /// Explicit concurrency value (None = auto-detect)
-    value: Option<usize>,
-    /// Whether to use auto-detection
-    auto_detect: bool,
-}
-
-impl Default for ConcurrencyConfig {
-    fn default() -> Self {
-        Self {
-            value: None,
-            auto_detect: true,
-        }
-    }
-}
-
-impl std::fmt::Display for ConcurrencyConfig {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if self.is_auto() {
-            write!(f, "auto")
-        } else if let Some(value) = self.value {
-            write!(f, "{}", value)
-        } else {
-            write!(f, "auto")
-        }
-    }
-}
-
-impl ConcurrencyConfig {
-    /// Create a new config with explicit value
-    ///
-    /// # Arguments
-    ///
-    /// * `value` - Explicit concurrency value (will be clamped 1-16)
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ConcurrencyConfig;
-    ///
-    /// let config = ConcurrencyConfig::new(5);
-    /// assert_eq!(config.resolve(), 5);
-    /// ```
-    #[must_use]
-    pub fn new(value: usize) -> Self {
-        Self {
-            value: Some(value.clamp(1, 16)),
-            auto_detect: false,
-        }
-    }
-
-    /// Create auto-detecting config (default)
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ConcurrencyConfig;
-    ///
-    /// let config = ConcurrencyConfig::auto();
-    /// let concurrency = config.resolve();
-    /// assert!(concurrency >= 1);
-    /// ```
-    #[must_use]
-    pub fn auto() -> Self {
-        Self::default()
-    }
-
-    /// Resolve the actual concurrency value
-    ///
-    /// Uses auto-detection based on CPU cores:
-    /// - 1-2 cores: 1 (avoid overwhelming system)
-    /// - 4 cores: 3 (HDD-aware default)
-    /// - 8+ cores: min(cores - 1, 8)
-    ///
-    /// # Returns
-    ///
-    /// Concurrency value between 1 and 16
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ConcurrencyConfig;
-    ///
-    /// // Explicit value
-    /// let config = ConcurrencyConfig::new(5);
-    /// assert_eq!(config.resolve(), 5);
-    ///
-    /// // Auto-detect
-    /// let config = ConcurrencyConfig::auto();
-    /// let value = config.resolve();
-    /// assert!(value >= 1 && value <= 16);
-    /// ```
-    pub fn resolve(&self) -> usize {
-        if let Some(value) = self.value {
-            return value;
-        }
-
-        // Auto-detect based on CPU cores
-        let cores = std::thread::available_parallelism()
-            .map(|p| p.get())
-            .unwrap_or(2);
-
-        // Smart defaults based on hardware
-        let optimal = match cores {
-            1 | 2 => 1,              // Single/dual-core: keep it simple
-            3 | 4 => 3,              // Quad-core: HDD-aware default
-            5..=7 => 5,              // 5-7 cores: good balance
-            _ => (cores - 1).min(8), // 8+ cores: cap at 8 for safety
-        };
-
-        optimal.clamp(1, 16)
-    }
-
-    /// Check if this config uses auto-detection
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rust_scraper::ConcurrencyConfig;
-    ///
-    /// let auto = ConcurrencyConfig::auto();
-    /// assert!(auto.is_auto());
-    ///
-    /// let explicit = ConcurrencyConfig::new(5);
-    /// assert!(!explicit.is_auto());
-    /// ```
-    #[must_use]
-    pub fn is_auto(&self) -> bool {
-        self.auto_detect && self.value.is_none()
-    }
-
-    /// Get the raw value if explicitly set
-    #[must_use]
-    pub fn get(&self) -> Option<usize> {
-        self.value
-    }
-}
-
-/// Custom value parser for clap (accepts "auto" or number)
-impl From<&str> for ConcurrencyConfig {
-    fn from(s: &str) -> Self {
-        let s = s.trim().to_lowercase();
-        if s == "auto" || s.is_empty() {
-            Self::default()
-        } else {
-            s.parse().map(ConcurrencyConfig::new).unwrap_or_else(|_| {
-                eprintln!("Warning: Invalid concurrency '{}', using auto-detect", s);
-                Self::default()
-            })
-        }
-    }
-}
-
-impl std::str::FromStr for ConcurrencyConfig {
-    type Err = std::num::ParseIntError;
-
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        let s = s.trim().to_lowercase();
-        if s == "auto" || s.is_empty() {
-            Ok(Self::default())
-        } else {
-            s.parse::<usize>().map(ConcurrencyConfig::new)
-        }
-    }
-}
-
-impl clap::builder::ValueParserFactory for ConcurrencyConfig {
-    type Parser = concurrency_parser::ConcurrencyValueParser;
-
-    fn value_parser() -> Self::Parser {
-        concurrency_parser::ConcurrencyValueParser
-    }
-}
-
-mod concurrency_parser {
-    use super::ConcurrencyConfig;
-    use clap::builder::TypedValueParser;
-
-    #[derive(Debug, Clone)]
-    pub struct ConcurrencyValueParser;
-
-    impl TypedValueParser for ConcurrencyValueParser {
-        type Value = ConcurrencyConfig;
-
-        fn parse_ref(
-            &self,
-            _cmd: &clap::Command,
-            _arg: Option<&clap::Arg>,
-            value: &std::ffi::OsStr,
-        ) -> Result<Self::Value, clap::Error> {
-            let value = value
-                .to_str()
-                .ok_or_else(|| clap::Error::new(clap::error::ErrorKind::InvalidUtf8))?;
-
-            let value = value.trim().to_lowercase();
-            if value.is_empty() || value == "auto" {
-                return Ok(ConcurrencyConfig::default());
-            }
-
-            value
-                .parse::<usize>()
-                .map(ConcurrencyConfig::new)
-                .map_err(|_| {
-                    clap::Error::raw(
-                        clap::error::ErrorKind::InvalidValue,
-                        format!(
-                            "'{}' is not a valid concurrency value (expected number or 'auto')",
-                            value
-                        ),
-                    )
-                })
-        }
-    }
-}
-
-/// CLI Arguments for the rust-scraper binary.
-///
-/// Parsed using `clap` with derive macros.
-///
-/// # Examples
-///
-/// ```no_run
-/// use rust_scraper::Args;
-/// use clap::Parser;
-///
-/// let args = Args::parse_from([
-///     "rust-scraper",
-///     "--url", "https://example.com",
-///     "--output", "./output",
-///     "--export-format", "jsonl",
-///     "--resume",
-/// ]);
-///
-/// assert_eq!(args.url, "https://example.com");
-/// ```
-#[derive(Parser, Debug)]
-#[command(name = "rust-scraper", version)]
-#[command(about = "Production-ready web scraper with Clean Architecture", long_about = None)]
-#[command(args_conflicts_with_subcommands = true)]
-pub struct Args {
-    /// Subcommands
-    #[command(subcommand)]
-    pub subcommand: Option<Commands>,
-
-    // ========== Target ==========
-    /// URL to scrape (required unless using a subcommand)
-    #[arg(short, long, env = "RUST_SCRAPER_URL")]
-    #[clap(next_help_heading = "Target")]
-    pub url: Option<String>,
-
-    /// CSS selector for content extraction
-    #[arg(short, long, default_value = "body", env = "RUST_SCRAPER_SELECTOR")]
-    #[clap(next_help_heading = "Target")]
-    pub selector: String,
-
-    // ========== Output ==========
-    /// Output directory for scraped content
-    #[arg(short, long, default_value = "output", env = "RUST_SCRAPER_OUTPUT")]
-    #[clap(next_help_heading = "Output")]
-    pub output: std::path::PathBuf,
-
-    /// Output format for individual files (markdown, text, json)
-    ///
-    /// Creates separate output files for each scraped page:
-    /// - markdown: Markdown with YAML frontmatter (default)
-    /// - text: Plain text without formatting
-    /// - json: Structured JSON with metadata
-    ///
-    /// Use this for human-readable output or when you need
-    /// individual files per page.
-    #[arg(
-        short = 'f',
-        long,
-        default_value = "markdown",
-        value_enum,
-        env = "RUST_SCRAPER_FORMAT"
-    )]
-    #[clap(next_help_heading = "Output")]
-    pub format: OutputFormat,
-
-    /// Export format for RAG pipeline (jsonl, vector, auto)
-    ///
-    /// Creates output suitable for retrieval-augmented generation:
-    /// - jsonl: JSON Lines format (one JSON per line), optimal for RAG
-    /// - vector: JSON format with metadata header and embeddings support
-    /// - auto: Detect from existing export files
-    ///
-    /// Use this for LLM/RAG pipelines that need batch export.
-    #[arg(
-        long,
-        default_value = "jsonl",
-        value_enum,
-        env = "RUST_SCRAPER_EXPORT_FORMAT"
-    )]
-    #[clap(next_help_heading = "Output")]
-    pub export_format: ExportFormat,
-
-    // ========== Obsidian Output ==========
-    /// Convert same-domain links to Obsidian [[wiki-link]] syntax
-    ///
-    /// Only links pointing to the same domain as the scraped page
-    /// are converted. External links remain as standard Markdown links.
-    ///
-    /// Example: [Read more](https://example.com/about) -> [[about|Read more]]
-    #[arg(
-        long,
-        default_value = "false",
-        env = "RUST_SCRAPER_OBSIDIAN_WIKI_LINKS"
-    )]
-    #[clap(next_help_heading = "Obsidian")]
-    pub obsidian_wiki_links: bool,
-
-    /// Tags to include in YAML frontmatter (comma-separated)
-    ///
-    /// Tags are added to the frontmatter of each Markdown file,
-    /// making them discoverable in Obsidian's tag pane.
-    ///
-    /// Example: --obsidian-tags "scraped,rust,web-dev"
-    #[arg(long, env = "RUST_SCRAPER_OBSIDIAN_TAGS", value_delimiter = ',')]
-    #[clap(next_help_heading = "Obsidian")]
-    pub obsidian_tags: Option<Vec<String>>,
-
-    /// Rewrite downloaded asset paths as relative to the .md file
-    ///
-    /// When images or documents are downloaded during scraping,
-    /// this flag rewrites their Markdown references to use relative
-    /// paths so they display correctly in Obsidian.
-    ///
-    /// Requires --download-images or --download-documents to have effect.
-    #[arg(
-        long,
-        default_value = "false",
-        env = "RUST_SCRAPER_OBSIDIAN_RELATIVE_ASSETS"
-    )]
-    #[clap(next_help_heading = "Obsidian")]
-    pub obsidian_relative_assets: bool,
-
-    /// Path to Obsidian vault (auto-detects if not provided)
-    ///
-    /// The vault must contain a `.obsidian/app.json` file.
-    /// If not provided, searches: OBSIDIAN_VAULT env var,
-    /// config file, then common locations (~/.obsidian, ~/Obsidian, etc.)
-    #[arg(long, env = "RUST_SCRAPER_OBSIDIAN_VAULT")]
-    #[clap(next_help_heading = "Obsidian")]
-    pub vault: Option<std::path::PathBuf>,
-
-    /// Quick-save mode: save directly to vault _inbox folder
-    ///
-    /// Bypasses TUI selection and saves all scraped content
-    /// to `<vault>/_inbox/`. Ideal for one-command clipping.
-    ///
-    /// Requires --obsidian-wiki-links to be enabled.
-    #[arg(
-        long,
-        default_value = "false",
-        env = "RUST_SCRAPER_OBSIDIAN_QUICK_SAVE"
-    )]
-    #[clap(next_help_heading = "Obsidian")]
-    pub quick_save: bool,
-
-    /// Add rich metadata to frontmatter (word count, reading time, language)
-    ///
-    /// Adds wordCount, readingTime, language, contentType, scrapeDate,
-    /// source, and status fields to YAML frontmatter for Dataview compatibility.
-    #[arg(
-        long,
-        default_value = "false",
-        env = "RUST_SCRAPER_OBSIDIAN_RICH_METADATA"
-    )]
-    #[clap(next_help_heading = "Obsidian")]
-    pub obsidian_rich_metadata: bool,
-
-    // ========== Discovery ==========
-    /// Delay between requests in milliseconds
-    #[arg(long, default_value = "1000", env = "RUST_SCRAPER_DELAY_MS")]
-    #[clap(next_help_heading = "Discovery")]
-    pub delay_ms: u64,
-
-    /// Maximum pages to scrape
-    #[arg(long, default_value = "10", env = "RUST_SCRAPER_MAX_PAGES")]
-    #[clap(next_help_heading = "Discovery")]
-    pub max_pages: usize,
-
-    // ========== Concurrency Control ==========
-    /// Concurrency level (number of parallel requests)
-    ///
-    /// Default: auto-detect based on CPU cores:
-    /// - 1-2 cores: 1
-    /// - 4 cores: 3 (HDD-aware)
-    /// - 8+ cores: min(CPU cores - 1, 8)
-    ///
-    /// Note: Can be overridden via CLI or detected at runtime.
-    /// The actual value used is determined at startup.
-    #[arg(long, default_value_t = ConcurrencyConfig::default(), env = "RUST_SCRAPER_CONCURRENCY")]
-    #[clap(next_help_heading = "Discovery")]
-    pub concurrency: ConcurrencyConfig,
-
-    // ========== Sitemap Support ==========
-    /// Use sitemap for URL discovery (auto-discovers from robots.txt if URL not provided)
-    #[arg(long, env = "RUST_SCRAPER_USE_SITEMAP")]
-    #[clap(next_help_heading = "Discovery")]
-    pub use_sitemap: bool,
-
-    /// Explicit sitemap URL (optional, auto-discovers if not provided)
-    #[arg(long, requires = "use_sitemap", env = "RUST_SCRAPER_SITEMAP_URL")]
-    #[clap(next_help_heading = "Discovery")]
-    pub sitemap_url: Option<String>,
-
-    // ========== Behavior ==========
-    /// Resume mode - skip URLs already processed
-    ///
-    /// Saves processing status to cache directory (~/.cache/rust-scraper/state)
-    /// Avoids re-processing URLs already scraped successfully.
-    #[arg(long, env = "RUST_SCRAPER_RESUME")]
-    #[clap(next_help_heading = "Behavior")]
-    pub resume: bool,
-
-    /// Custom state directory for resume mode
-    ///
-    /// Default: ~/.cache/rust-scraper/state
-    #[arg(long, env = "RUST_SCRAPER_STATE_DIR")]
-    #[clap(next_help_heading = "Behavior")]
-    pub state_dir: Option<std::path::PathBuf>,
-
-    /// Download images from the page
-    #[arg(long, default_value = "false", env = "RUST_SCRAPER_DOWNLOAD_IMAGES")]
-    #[clap(next_help_heading = "Behavior")]
-    pub download_images: bool,
-
-    /// Download documents from the page (PDF, DOCX, XLSX, etc.)
-    #[arg(long, default_value = "false", env = "RUST_SCRAPER_DOWNLOAD_DOCUMENTS")]
-    #[clap(next_help_heading = "Behavior")]
-    pub download_documents: bool,
-
-    // ========== TUI Interactive Mode ==========
-    /// Interactive mode with TUI URL selector
-    #[arg(long, env = "RUST_SCRAPER_INTERACTIVE")]
-    #[clap(next_help_heading = "Behavior")]
-    pub interactive: bool,
-
-    // ========== AI Semantic Cleaning ==========
-    /// Use AI-powered semantic cleaning for better RAG output
-    ///
-    /// When enabled:
-    /// - Uses SemanticCleaner to process HTML content
-    /// - Generates semantic chunks with embeddings
-    /// - Exports in JSONL format with embeddings field
-    ///
-    /// Requires: --features ai to be enabled at compile time
-    ///
-    /// Example:
-    ///   cargo run --features ai -- --url https://example.com --clean-ai
-    #[cfg(feature = "ai")]
-    #[arg(long, default_value = "false", env = "RUST_SCRAPER_CLEAN_AI")]
-    #[clap(next_help_heading = "Behavior")]
-    pub clean_ai: bool,
-
-    /// Feature flag placeholder when AI is not enabled
-    ///
-    /// This field exists to provide a clear error message when --clean-ai
-    /// is used without the 'ai' feature enabled.
-    #[cfg(not(feature = "ai"))]
-    #[arg(
-        long,
-        default_value = "false",
-        hide = true,
-        env = "RUST_SCRAPER_CLEAN_AI"
-    )]
-    pub clean_ai: bool,
-
-    // ========== JavaScript Rendering (reserved for v1.4) ==========
-    /// Force JavaScript rendering for SPA sites (not yet implemented)
-    ///
-    /// Reserved for future use. When implemented, this will enable
-    /// headless browser rendering for sites that require JavaScript
-    /// to display content (Single Page Applications).
-    ///
-    /// Currently this flag is a no-op and has no effect.
-    /// Track implementation: https://github.com/XaviCode1000/rust-scraper/issues/16
-    #[arg(long, default_value = "false", env = "RUST_SCRAPER_FORCE_JS_RENDER")]
-    #[clap(next_help_heading = "Behavior")]
-    pub force_js_render: bool,
-
-    // ========== Display ==========
-    /// Verbosity level (use multiple times for more detail: -v, -vv, -vvv)
-    #[arg(short, long, action = clap::ArgAction::Count, env = "RUST_SCRAPER_VERBOSE")]
-    #[clap(next_help_heading = "Display")]
-    pub verbose: u8,
-
-    /// T-020: Quiet mode — suppress info/debug output, show warnings only
-    #[arg(short = 'q', long, default_value = "false", env = "RUST_SCRAPER_QUIET")]
-    #[clap(next_help_heading = "Display")]
-    pub quiet: bool,
-
-    /// T-021: Dry-run mode — discover URLs and print them without scraping
-    #[arg(
-        short = 'n',
-        long,
-        default_value = "false",
-        env = "RUST_SCRAPER_DRY_RUN"
-    )]
-    #[clap(next_help_heading = "Display")]
-    pub dry_run: bool,
-
-    // ========== Crawler Settings ==========
-    /// Maximum depth to crawl (0 = only seed URL)
-    #[arg(long, default_value = "2", env = "RUST_SCRAPER_MAX_DEPTH")]
-    #[clap(next_help_heading = "Crawler Settings")]
-    pub max_depth: u8,
-
-    /// Request timeout in seconds
-    #[arg(long, default_value = "30", env = "RUST_SCRAPER_TIMEOUT_SECS")]
-    #[clap(next_help_heading = "Crawler Settings")]
-    pub timeout_secs: u64,
-
-    /// URL patterns to include (glob-style, can be repeated)
-    #[arg(
-        long = "include-pattern",
-        env = "RUST_SCRAPER_INCLUDE",
-        value_delimiter = ','
-    )]
-    #[clap(next_help_heading = "Crawler Settings")]
-    pub include_patterns: Vec<String>,
-
-    /// URL patterns to exclude (glob-style, can be repeated)
-    #[arg(
-        long = "exclude-pattern",
-        env = "RUST_SCRAPER_EXCLUDE",
-        value_delimiter = ','
-    )]
-    #[clap(next_help_heading = "Crawler Settings")]
-    pub exclude_patterns: Vec<String>,
-
-    // ========== HTTP Client Settings ==========
-    /// Maximum number of retry attempts for failed requests
-    #[arg(long, default_value = "3", env = "RUST_SCRAPER_MAX_RETRIES")]
-    #[clap(next_help_heading = "HTTP Client Settings")]
-    pub max_retries: u32,
-
-    /// Base delay for exponential backoff in milliseconds
-    #[arg(long, default_value = "1000", env = "RUST_SCRAPER_BACKOFF_BASE_MS")]
-    #[clap(next_help_heading = "HTTP Client Settings")]
-    pub backoff_base_ms: u64,
-
-    /// Maximum delay for exponential backoff in milliseconds
-    #[arg(long, default_value = "10000", env = "RUST_SCRAPER_BACKOFF_MAX_MS")]
-    #[clap(next_help_heading = "HTTP Client Settings")]
-    pub backoff_max_ms: u64,
-
-    /// Accept-Language header value
-    #[arg(
-        long,
-        default_value = "en-US,en;q=0.9",
-        env = "RUST_SCRAPER_ACCEPT_LANGUAGE"
-    )]
-    #[clap(next_help_heading = "HTTP Client Settings")]
-    pub accept_language: String,
-
-    // ========== Download Settings ==========
-    /// Maximum file size to download in bytes (default: 50MB)
-    #[arg(long, default_value = "52428800", env = "RUST_SCRAPER_MAX_FILE_SIZE")]
-    #[clap(next_help_heading = "Download Settings")]
-    pub max_file_size: u64,
-
-    /// Timeout for individual asset downloads in seconds
-    #[arg(long, default_value = "30", env = "RUST_SCRAPER_DOWNLOAD_TIMEOUT")]
-    #[clap(next_help_heading = "Download Settings")]
-    pub download_timeout: u64,
-
-    // ========== AI Settings (feature-gated) ==========
-    /// Relevance threshold for AI semantic filtering (0.0-1.0)
-    #[cfg(feature = "ai")]
-    #[arg(long, default_value = "0.3", env = "RUST_SCRAPER_THRESHOLD")]
-    #[clap(next_help_heading = "AI Settings")]
-    pub threshold: f32,
-
-    /// Maximum tokens per chunk for AI processing
-    #[cfg(feature = "ai")]
-    #[arg(long, default_value = "512", env = "RUST_SCRAPER_MAX_TOKENS")]
-    #[clap(next_help_heading = "AI Settings")]
-    pub max_tokens: usize,
-
-    /// Run AI model in offline mode (fail if not cached)
-    #[cfg(feature = "ai")]
-    #[arg(long, default_value = "false", env = "RUST_SCRAPER_OFFLINE", action = clap::ArgAction::SetTrue)]
-    #[clap(next_help_heading = "AI Settings")]
-    pub offline: bool,
-
-    // ========== Sitemap Settings ==========
-    /// Maximum recursion depth for sitemap indexes
-    #[arg(long, default_value = "3", env = "RUST_SCRAPER_SITEMAP_DEPTH")]
-    #[clap(next_help_heading = "Sitemap Settings")]
-    pub sitemap_depth: u8,
-}
-
-// T-024: Version string with build metadata
 /// Return the extended version string including git commit and build date.
 pub fn version_string() -> String {
     let commit = built_info::GIT_COMMIT_HASH_SHORT.unwrap_or("unknown");
@@ -1048,249 +142,4 @@ pub fn version_string() -> String {
         commit,
         build
     )
-}
-
-/// Subcommands (T-041: completions subcommand)
-#[derive(Debug, clap::Subcommand)]
-pub enum Commands {
-    /// Generate shell completion scripts
-    Completions {
-        /// Shell to generate completions for (bash, elvish, fish, powershell, zsh)
-        #[arg(value_enum)]
-        shell: Shell,
-    },
-}
-
-/// Shell type for completions, re-exported from clap_complete
-#[derive(Debug, Clone, Copy, PartialEq, Eq, clap::ValueEnum)]
-pub enum Shell {
-    Bash,
-    Elvish,
-    Fish,
-    PowerShell,
-    Zsh,
-}
-
-impl From<Shell> for clap_complete::Shell {
-    fn from(s: Shell) -> Self {
-        match s {
-            Shell::Bash => clap_complete::Shell::Bash,
-            Shell::Elvish => clap_complete::Shell::Elvish,
-            Shell::Fish => clap_complete::Shell::Fish,
-            Shell::PowerShell => clap_complete::Shell::PowerShell,
-            Shell::Zsh => clap_complete::Shell::Zsh,
-        }
-    }
-}
-
-// ============================================================================
-// Public Functions
-// ============================================================================
-
-/// Validate and parse a URL string using the `url` crate (RFC 3986 compliant).
-///
-/// This function performs strict URL validation:
-/// - Trims whitespace automatically
-/// - Requires http or https scheme (case-insensitive)
-/// - Requires a valid host
-/// - Rejects malformed URLs
-///
-/// # Arguments
-///
-/// * `url` - URL string to validate and parse
-///
-/// # Returns
-///
-/// * `Ok(url::Url)` - Validated and parsed URL
-/// * `Err(ScraperError::InvalidUrl)` - Invalid URL with error message
-///
-/// # Errors
-///
-/// Returns an error if:
-/// - URL is empty
-/// - URL has invalid format
-/// - URL scheme is not http or https
-/// - URL has no host
-///
-/// # Examples
-///
-/// ```
-/// use rust_scraper::validate_and_parse_url;
-///
-/// // Valid URLs
-/// let url = validate_and_parse_url("https://example.com").unwrap();
-/// assert_eq!(url.host_str(), Some("example.com"));
-///
-/// let url = validate_and_parse_url("HTTP://EXAMPLE.COM").unwrap();
-/// assert_eq!(url.scheme(), "http");
-///
-/// // Invalid URLs
-/// assert!(validate_and_parse_url("").is_err());
-/// assert!(validate_and_parse_url("ftp://example.com").is_err());
-/// assert!(validate_and_parse_url("not-a-url").is_err());
-/// ```
-///
-/// # Whitespace Handling
-///
-/// Leading and trailing whitespace is automatically trimmed:
-///
-/// ```
-/// use rust_scraper::validate_and_parse_url;
-///
-/// let url = validate_and_parse_url("  https://example.com  ").unwrap();
-/// assert_eq!(url.host_str(), Some("example.com"));
-/// ```
-pub fn validate_and_parse_url(url: &str) -> Result<url::Url> {
-    if url.is_empty() {
-        return Err(ScraperError::invalid_url("URL cannot be empty"));
-    }
-
-    // Url::parse automatically trims whitespace and handles case-insensitive schemes
-    // Following rust-skills: url-no-string-split (don't use starts_with for URLs)
-    let parsed = url::Url::parse(url.trim())
-        .map_err(|e| ScraperError::invalid_url(format!("Failed to parse URL '{}': {}", url, e)))?;
-
-    // Check scheme (case-insensitive, already lowercased by Url::parse)
-    match parsed.scheme() {
-        "http" | "https" => {},
-        scheme => {
-            return Err(ScraperError::invalid_url(format!(
-                "URL must use http or https scheme, got '{}'",
-                scheme
-            )))
-        },
-    }
-
-    if parsed.host_str().is_none() {
-        return Err(ScraperError::invalid_url("URL must have a valid host"));
-    }
-
-    Ok(parsed)
-}
-
-// ============================================================================
-// Tests
-// ============================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_scraper_config_default() {
-        let config = ScraperConfig::default();
-        assert!(!config.download_images);
-        assert!(!config.download_documents);
-        assert!(!config.has_downloads());
-        assert_eq!(config.scraper_concurrency, 3);
-    }
-
-    #[test]
-    fn test_scraper_config_with_images() {
-        let config = ScraperConfig::default().with_images();
-        assert!(config.download_images);
-        assert!(config.has_downloads());
-    }
-
-    #[test]
-    fn test_scraper_config_with_documents() {
-        let config = ScraperConfig::default().with_documents();
-        assert!(config.download_documents);
-        assert!(config.has_downloads());
-    }
-
-    #[test]
-    fn test_scraper_config_with_concurrency() {
-        let config = ScraperConfig::default().with_scraper_concurrency(5);
-        assert_eq!(config.scraper_concurrency, 5);
-    }
-
-    #[test]
-    fn test_validate_and_parse_url_success() {
-        let url = validate_and_parse_url("https://example.com");
-        assert!(url.is_ok());
-    }
-
-    #[test]
-    fn test_validate_and_parse_url_empty() {
-        let result = validate_and_parse_url("");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_validate_and_parse_url_invalid_scheme() {
-        let result = validate_and_parse_url("ftp://example.com");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_validate_and_parse_url_whitespace() {
-        let url = validate_and_parse_url("  https://example.com  ");
-        assert!(url.is_ok());
-        assert_eq!(url.unwrap().host_str(), Some("example.com"));
-    }
-
-    #[test]
-    fn test_concurrency_config_new() {
-        let config = ConcurrencyConfig::new(5);
-        assert_eq!(config.resolve(), 5);
-    }
-
-    #[test]
-    fn test_concurrency_config_auto() {
-        let config = ConcurrencyConfig::auto();
-        let value = config.resolve();
-        assert!((1..=16).contains(&value));
-    }
-
-    #[test]
-    fn test_concurrency_config_clamp() {
-        let config = ConcurrencyConfig::new(100);
-        assert_eq!(config.resolve(), 16); // Clamped to max
-    }
-
-    #[test]
-    fn test_concurrency_config_display() {
-        let auto = ConcurrencyConfig::auto();
-        assert_eq!(format!("{}", auto), "auto");
-
-        let explicit = ConcurrencyConfig::new(5);
-        assert_eq!(format!("{}", explicit), "5");
-    }
-
-    #[test]
-    fn test_concurrency_config_from_str() {
-        let config = ConcurrencyConfig::from("5");
-        assert_eq!(config.resolve(), 5);
-
-        let config = ConcurrencyConfig::from("auto");
-        assert!(config.is_auto());
-
-        let config = ConcurrencyConfig::from("");
-        assert!(config.is_auto());
-    }
-
-    #[test]
-    fn test_concurrency_config_from_str_invalid() {
-        // Should fallback to auto with warning (tested via output)
-        let config = ConcurrencyConfig::from("not-a-number");
-        assert!(config.is_auto());
-    }
-
-    #[test]
-    fn test_export_format_from_str() {
-        // Test ExportFormat parsing
-        let format = ExportFormat::parse_str("jsonl");
-        assert!(format.is_ok());
-    }
-
-    // ============================================================================
-    // Tests for Args CLI arguments
-    // NOTE: These tests use clap's test utilities which have a known issue in clap 4.6
-    // when subcommands are specified. The Args struct functionality is tested in
-    // tests/integration_test.rs::test_args_has_required_fields instead.
-    // ============================================================================
-
-    // NOTE: Tests disabled due to clap 4.6 debug_assert issue with subcommands in test context
-    // The actual functionality is verified by integration_test.rs
 }
